+++ pypeit/specobjs.py
@@ -804,6 +804,9 @@ class SpecObjs:
                 opt_fwhm.append(np.median(specobj.FWHMFIT) * binspatial * platescale)
             else:  # Optimal is not required to occur
                 opt_fwhm.append(0.)
+            # NOTE: Below requires that S2N not be None, otherwise the code will
+            # fault.  If the code gets here and S2N is None, check that 1D
+            # extractions have been performed.
             s2n.append(specobj.S2N)
             # Manual extraction?
             manual_extract.append(specobj.hand_extract_flag)